+++ lib/info.dart
@@ -2,6 +2,6 @@ class INFO {
   static const String title = 'IRIS';
   static const String author = '22';
   static const String authorUrl = 'https://github.com/nini22P';
-  static const String githubUrl = 'https://github.com/nini22P/iris';
+  static const String githubUrl = 'https://github.com/U707T/iris';
   static const String msStoreId = '9NML7WNHNRTJ';
 }